Address media test review comments
